chore: update page meta
