+++ themes/theme-default/src/client/components/VPPageMeta.vue
@@ -1,90 +1,11 @@
 <script setup lang="ts">
-import { computed } from 'vue'
-import type { ComputedRef } from 'vue'
-import { AutoLink, usePageData, usePageFrontmatter } from 'vuepress/client'
-import type { AutoLinkConfig } from 'vuepress/client'
-import type {
-  DefaultThemeNormalPageFrontmatter,
-  DefaultThemePageData,
-} from '../../shared/index.js'
-import { useThemeLocaleData } from '../composables/index.js'
-import { resolveEditLink } from '../utils/index.js'
-
-const useEditNavLink = (): ComputedRef<null | AutoLinkConfig> => {
-  const themeLocale = useThemeLocaleData()
-  const page = usePageData<DefaultThemePageData>()
-  const frontmatter = usePageFrontmatter<DefaultThemeNormalPageFrontmatter>()
-
-  return computed(() => {
-    const showEditLink =
-      frontmatter.value.editLink ?? themeLocale.value.editLink ?? true
-    if (!showEditLink) {
-      return null
-    }
-
-    const {
-      repo,
-      docsRepo = repo,
-      docsBranch = 'main',
-      docsDir = '',
-      editLinkText,
-    } = themeLocale.value
-
-    if (!docsRepo) return null
-
-    const editLink = resolveEditLink({
-      docsRepo,
-      docsBranch,
-      docsDir,
-      filePathRelative: page.value.filePathRelative,
-      editLinkPattern:
-        frontmatter.value.editLinkPattern ?? themeLocale.value.editLinkPattern,
-    })
-
-    if (!editLink) return null
-
-    return {
-      text: editLinkText ?? 'Edit this page',
-      link: editLink,
-    }
-  })
-}
-
-const useLastUpdated = (): ComputedRef<null | string> => {
-  const themeLocale = useThemeLocaleData()
-  const page = usePageData<DefaultThemePageData>()
-  const frontmatter = usePageFrontmatter<DefaultThemeNormalPageFrontmatter>()
-
-  return computed(() => {
-    const showLastUpdated =
-      frontmatter.value.lastUpdated ?? themeLocale.value.lastUpdated ?? true
-
-    if (!showLastUpdated) return null
-
-    if (!page.value.git?.updatedTime) return null
-
-    const updatedDate = new Date(page.value.git?.updatedTime)
-
-    return updatedDate.toLocaleString()
-  })
-}
-
-const useContributors = (): ComputedRef<
-  null | Required<DefaultThemePageData['git']>['contributors']
-> => {
-  const themeLocale = useThemeLocaleData()
-  const page = usePageData<DefaultThemePageData>()
-  const frontmatter = usePageFrontmatter<DefaultThemeNormalPageFrontmatter>()
-
-  return computed(() => {
-    const showContributors =
-      frontmatter.value.contributors ?? themeLocale.value.contributors ?? true
-
-    if (!showContributors) return null
-
-    return page.value.git?.contributors ?? null
-  })
-}
+import { AutoLink } from 'vuepress/client'
+import {
+  useContributors,
+  useEditNavLink,
+  useLastUpdated,
+  useThemeLocaleData,
+} from '../composables/index.js'
 
 const themeLocale = useThemeLocaleData()
 const editNavLink = useEditNavLink()
