+++ themes/theme-default/src/client/composables/index.ts
@@ -1,4 +1,7 @@
+export * from './useContributors.js'
 export * from './useDarkMode.js'
+export * from './useEditLink.js'
+export * from './useLastUpdated.js'
 export * from './useNavigate.js'
 export * from './useNavbarConfig.js'
 export * from './useNavbarRepo.js'
